Move legacy SimpleToast files
