+++ components/simple-toast/legacy-styled.jsx
